added angular universal

+++ src/app/core/auth/services/auth.service.ts
@@ -1,8 +1,9 @@
-import { Injectable } from '@angular/core';
+import { Inject, Injectable, PLATFORM_ID } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable, BehaviorSubject, Subject } from 'rxjs';
 import { environment } from '../../../../environments/environment';
 import { Router } from '@angular/router';
+import { isPlatformBrowser } from '@angular/common';
 
 @Injectable({
   providedIn: 'root',
@@ -15,32 +16,34 @@ export class AuthService {
   public isLogoutState = this.isLogoutSubject.asObservable();
   public isUserOperationSource = new BehaviorSubject<boolean>(false);
   public isUserOperationState = this.isUserOperationSource.asObservable();
-  constructor(private http: HttpClient, private router: Router) {}
+  constructor(
+    private http: HttpClient,
+    private router: Router,
+    @Inject(PLATFORM_ID) private platformId: Object
+  ) {}
 
   public saveToken(token: string): void {
-    localStorage.setItem('token', token);
-    this.tokenSubjectSource.next(token);
-    this.token = token;
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.setItem('token', token);
+      this.tokenSubjectSource.next(token);
+      this.token = token;
+    }
   }
 
   public getToken(): string {
-    if (!this.token) {
+    if (!this.token && isPlatformBrowser(this.platformId)) {
       let token = localStorage.getItem('token') || '';
       this.token = token;
     }
     return this.token;
   }
 
-  public getUserIdWhenLoginIn(): string {
-    return localStorage.getItem('userId');
-  }
-
-
-
   public logOut() {
-    this.token = null;
-    localStorage.clear();
-    this.router.navigate([`/auth/login`]);
+    if (isPlatformBrowser(this.platformId)) {
+      this.token = null;
+      localStorage.clear();
+      this.router.navigate([`/auth/login`]);
+    }
   }
 
   public login($userCredentials: any): Observable<any> {
